converting either UV or scale from half to float will fix the close moire pattern. changing scale to have only one dimension converted and reduce number of floats needed.

+++ Assets/Glass/Glass.shader
@@ -251,7 +251,7 @@
                 return t;
             }
 
-            half SimpleNoise(half2 UV, half Scale)
+            half SimpleNoise(half2 UV, float Scale)
             {
                 half t = 0.0;
